refactor: consolidate type generation into adapter
Move type generation logic into `Adapter.generateTypes()` to eliminate
duplication across init and sync commands, and remove the now-unused
`src/lib/codegen.ts` module.
